Split ndc-models lib.rs
